// src/types/dataverse.ts
export interface PsDeliveryroutes {
  ps_deliveryroutesid?: string
  createdby?: string | null
  createdbyname?: string | null
  createdbyyominame?: string | null
  createdon?: string | null
  createdonbehalfby?: string | null
  createdonbehalfbyname?: string | null
  createdonbehalfbyyominame?: string | null
  importsequencenumber?: number | null
  modifiedby?: string | null
  modifiedbyname?: string | null
  modifiedbyyominame?: string | null
  modifiedon?: string | null
  modifiedonbehalfby?: string | null
  modifiedonbehalfbyname?: string | null
  modifiedonbehalfbyyominame?: string | null
  overriddencreatedon?: string | null
  ownerid?: string | null
  owneridname?: string | null
  owneridtype?: string | null
  owneridyominame?: string | null
  owningbusinessunit?: string | null
  owningbusinessunitname?: string | null
  owningteam?: string | null
  owninguser?: string | null
  ps_sequence?: number | null
  ps_sitelat?: number | null
  ps_sitelong?: number | null
  ps_vehiclename?: string | null
  statecode?: number | null
  statecodename?: unknown
  statuscode?: number | null
  statuscodename?: unknown
  timezoneruleversionnumber?: number | null
  utcconversiontimezonecode?: number | null
  versionnumber?: number | null
}
// ...
export type SalesorderEntitySet = 'salesorders'
// ...
export const salesorderEntitySet: SalesorderEntitySet = 'salesorders'